lab_pneumonia/fastapi_deployment/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
numpy==1.23.5
Pillow==9.5.0
keras==2.12.0
tensorflow==2.12.0

